// src/app/enums/enumAlertText.ts
export enum enumAlertText {

  formTitle = 'Formulario',
  noAuthorized = 'No autorizado',
  errorTitle = 'Error',
  warningTitle = 'Atención',
  passwordTitle = 'Contraseña',
  // Descriptions
  errorForm= 'Complete el formulario correctamente.',
  errorSubmit= 'Ocurrió un error durante el proceso, intente de nuevo.',
  errorGet ='Ocurrio un error al momento de obtener los datos.',
  requiredFields = 'Completá los campos vacíos.',
  passwordNotValid = 'La contraseña debe ser mayor a 8 caracteres y debe contener al menos una letra mayúscula.',
  passwordNotEqual = 'Las contraseñas no coinciden.',
  invalidRoute = 'La URL que solicitaste es incorrecta',
  formSaveSuccess = 'Se ha guardo correctamente',
  formUpdateSuccess = 'Se ha actualizado correctamente',
  formDeleteSuccess = 'Se ha eliminado correctamente',
}